+++ BuildScripts/jprm_build.sh
@@ -67,10 +67,14 @@ rc=$?
 # package Templates/ as well
 cd ${MY}/../${JELLYFIN_REPO}
 zipfile="./newsletters/newsletters_${VERSION}.zip"
+oldChecksum=$(md5sum "${zipfile}" | awk '{print $1}')
+echo "Old checksum:: ${oldChecksum}"
 zip -r ${zipfile} ./Templates
 echo "----------"
 echo "Contents in ${zipfile}"
 unzip -l ${zipfile}
+newChecksum=$(md5sum "${zipfile}" | awk '{print $1}')
+echo "New checksum:: ${newChecksum}"
 sed -i "s/github.com\/solidsnake1298\/Jellyfin-Newsletter-Plugin\/releases\/download\/newsletters/github.com\/solidsnake1298\/Jellyfin-Newsletter-Plugin\/releases\/download\/v${VERSION}/g" manifest.json
-md5sum ${zipfile}
+sed -i "s/$oldChecksum/$newChecksum/g" manifest.json
 exit $rc