+++ Jellyfin.Plugin.Newsletters/Shared/Entities/JsonFileObj.cs
@@ -12,7 +12,7 @@ namespace Jellyfin.Plugin.Newsletters.Scripts.ENTITIES;
 public class JsonFileObj
 {
     private Logger? logger;
-
+    
     /// <summary>
     /// Initializes a new instance of the <see cref="JsonFileObj"/> class.
     /// </summary>
@@ -24,7 +24,6 @@ public class JsonFileObj
         Season = 0;
         Episode = 0;
         Overview = string.Empty;
-        ImageURL = string.Empty;
         ItemID = string.Empty;
         PosterPath = string.Empty;
         Type = string.Empty;
@@ -43,8 +42,6 @@ public class JsonFileObj
 
     public string Overview { get; set; }
 
-    public string ImageURL { get; set; }
-
     public string ItemID { get; set; }
 
     public string PosterPath { get; set; }
@@ -55,18 +52,6 @@ public class JsonFileObj
 
     public JsonFileObj ConvertToObj(IReadOnlyList<ResultSetValue> row)
     {
-        // Filename = string.Empty; 0
-        // Title = string.Empty; 1
-        // Album = string.Empty; 2
-        // Season = 0; 3
-        // Episode = 0; 4
-        // Overview = string.Empty; 5
-        // ImageURL = string.Empty; 6
-        // ItemID = string.Empty; 7
-        // PosterPath = string.Empty; 8
-        // Album = string.Empty; 9
-        // Emailed = 0; 10
-
         logger = new Logger();
         JsonFileObj obj = new JsonFileObj()
         {
@@ -76,11 +61,10 @@ public class JsonFileObj
             Season = int.Parse(row[3].ToString(), CultureInfo.CurrentCulture),
             Episode = int.Parse(row[4].ToString(), CultureInfo.CurrentCulture),
             Overview = row[5].ToString(),
-            ImageURL = row[6].ToString(),
-            ItemID = row[7].ToString(),
-            PosterPath = row[8].ToString(),
-            Type = row[9].ToString(),
-            Emailed = int.Parse(row[10].ToString(), CultureInfo.CurrentCulture)
+            ItemID = row[6].ToString(),
+            PosterPath = row[7].ToString(),
+            Type = row[8].ToString(),
+            Emailed = int.Parse(row[9].ToString(), CultureInfo.CurrentCulture)
         };
 
         return obj;
@@ -95,10 +79,10 @@ public class JsonFileObj
         item_dict.Add("{Season}", this.Season);
         item_dict.Add("{Episode}", this.Episode);
         item_dict.Add("{Overview}", this.Overview);
-        item_dict.Add("{ImageURL}", this.ImageURL);
         item_dict.Add("{ItemID}", this.ItemID);
         item_dict.Add("{PosterPath}", this.PosterPath);
         item_dict.Add("{Type}", this.Type);
+        item_dict.Add("{ImageURL}", "cid:" + this.ItemID);
 
         return item_dict;        
     }